# go-client-register-go/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Get apps endpoints from eureka"
  version: "1.0.0"
  title: "Register-go API"
host: "localhost:1906"
basePath: "/api"
tags:
- name: "apps"
  description: "Info about apps and containers"
schemes:
- "http"
paths:
  /apps/{appName}:
    get:
      tags:
      - "apps"
      summary: "Get an app endpoint by app name"
      description: "Returns an app endpoint"
      operationId: "GetAppsByName"
      produces:
      - "application/json"
      parameters:
      - name: "appName"
        in: "path"
        description: "App name"
        required: true
        type: "string"
        x-exportParamName: "AppName"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/App"
        404:
          description: "App not found"
  /apps/{appName}/all:
    get:
      tags:
      - "apps"
      summary: "Get all apps endpoints by app name"
      description: "Returns an array of app endpoint"
      operationId: "GetAllAppsByName"
      produces:
      - "application/json"
      parameters:
      - name: "appName"
        in: "path"
        description: "App name"
        required: true
        type: "string"
        x-exportParamName: "AppName"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/App"
        404:
          description: "App not found"
  /register:
    get:
      tags:
      - "apps"
      summary: "Register an app on Eureka"
      description: "Returns success"
      operationId: "Register"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Msg"
definitions:
  App:
    type: "object"
    properties:
      instanceId:
        type: "string"
        example: "app_hostname_port"
      endpoint:
        type: "string"
        example: "http://hostname:port"
    example:
      endpoint: "http://hostname:port"
      instanceId: "app_hostname_port"
  Msg:
    type: "object"
    properties:
      content:
        type: "string"
        example: "success"
    example:
      content: "success"