# go-client-register-go/app.go
/*
 * Register-go API
 *
 * Get apps endpoints from eureka
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type App struct {

	InstanceId string `json:"instanceId,omitempty"`

	Endpoint string `json:"endpoint,omitempty"`
}